use "size_t" instead of "int" as default idx_type

+++ chunky_mem/include/chunky_mem/chunk.h
@@ -12,7 +12,7 @@
 namespace chunky_mem {
 
     template<
-        typename idx_t = int,
+        typename idx_t = size_t,
         typename size_t = size_t
     >
     struct Chunk
